// src/lib/glx/data/ops.gen.ts
export type GenPickup = { pickup_id: string; order_id: string; attempt_no: number; result: string; reason: string; attempt_ts: string };
export const GEN_PICKUPS: GenPickup[] = [
  {
    "pickup_id": "GLX-PCK-G005-1",
    "order_id": "GLX-ORD-G005",
    "attempt_no": 1,
    "result": "failed",
    "reason": "customer_not_home",
    "attempt_ts": "2026-08-12T11:00:00"
  },
  {
    "pickup_id": "GLX-PCK-G005-2",
    "order_id": "GLX-ORD-G005",
    "attempt_no": 2,
    "result": "failed",
    "reason": "customer_not_home",
    "attempt_ts": "2026-08-14T11:00:00"
  }
];